+++ cpp/order/Wigner3j.h
@@ -7,8 +7,6 @@
 #include <complex>
 #include <vector>
 
-using namespace std;
-
 /*! \file Wigner3j.h
  *  \brief Stores and reduces over Wigner 3j coefficients for l from 0 to 20
  */
@@ -22,9 +20,10 @@ int lmIndex(int l, int m);
 //! Reduce an array using Wigner 3j coefficients to construct a
 //  third-order rotational invariant quantity.
 //  source array must be indexed by m, like [0, 1, ..., l, -1, -2, ..., -l].
-float reduceWigner3j(const complex<float>* source, unsigned int l_, const vector<double> &wigner3j);
+float reduceWigner3j(const std::complex<float>* source,
+        unsigned int l_, const std::vector<double> &wigner3j);
 
-vector<double> getWigner3j(unsigned int l);
+std::vector<double> getWigner3j(unsigned int l);
 // All Wigner 3j coefficients created using sympy
 /*