+++ freud/cluster.pyx
@@ -161,12 +161,12 @@ cdef class Cluster(PairCompute):
         """
         import freud.plot
         try:
-            count = np.unique(self.cluster_idx, return_counts=True)
+            values, counts = np.unique(self.cluster_idx, return_counts=True)
         except ValueError:
             return None
         else:
-            return freud.plot.clusters_plot(count[0], count[1],
-                                            num_clusters_to_plot=10, ax=ax)
+            return freud.plot.clusters_plot(
+                values, counts, num_clusters_to_plot=10, ax=ax)
 
     def _repr_png_(self):
         import freud.plot